Split global args into PairedToken args and DiffString args
